// Frontend/src/types/client.ts
export type clientRequest = {
  name: string;
  businessNumber: string;
  address: string;
  tel: string;
  email: string;
  representativeName: string;
};

export type clientResponse = {
  id: number;
  name: string;
  businessNumber: string;
  address: string;
  tel: string;
  email: string;
  representativeName: string;
};